# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.Project:
    properties:
      color:
        type: string
      created_at:
        description: omitempty, чтобы не возвращать null
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      owner_id:
        type: integer
      parent_project_id:
        type: integer
      updated_at:
        description: omitempty, чтобы не возвращать null
        type: string
    type: object
  dto.ProjectUser:
    properties:
      invited_at:
        type: string
      invited_by_user_id:
        type: integer
      joined_at:
        type: string
      permission:
        type: string
      project_id:
        type: integer
      user_email:
        type: string
      user_id:
        type: integer
    type: object
  dto.ProjectUserInvite:
    properties:
      permission:
        type: string
      project_id:
        type: integer
      user_email:
        type: string
    type: object
...
  project_handler.HandleCreateProjectParam:
    properties:
      color:
        type: string
      description:
        type: string
      name:
        type: string
      parent_project_id:
        type: integer
    required:
    - description
    - name
    type: object
  response.JSONResponse:
    description: default JSON-response
    properties:
      data:
        type: object
      error:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Task manager
  title: DataTask
  version: "1.0"
paths:
...
  /project:
    post:
      consumes:
      - application/json
      description: Create a new project
      parameters:
      - description: Project data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project_handler.HandleCreateProjectParam'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Create Project
      tags:
      - Project
  /project/{id}:
    delete:
      description: Delete a project by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Delete Project
      tags:
      - Project
    get:
      description: Get a project by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Project by ID
      tags:
      - Project
    put:
      consumes:
      - application/json
      description: Update a project's details
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated project data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.Project'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Update Project
      tags:
      - Project
  /project_subprojects/{parent_project_id}:
    get:
      description: Get all subprojects of a project
      parameters:
      - description: Parent Project ID
        in: path
        name: parent_project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Project'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Subprojects
      tags:
      - Project
...
  /project_users/{project_id}:
    get:
      description: Get all users and their permissions in a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ProjectUser'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Users in Project
      tags:
      - Project
  /project_users/{project_id}/accept:
    post:
      description: Accept an invitation to join a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Accept Project Invitation
      tags:
      - Project
  /project_users/{project_id}/invite:
    post:
      consumes:
      - application/json
      description: Invite a user to a project with specific permissions
      parameters:
      - description: User invitation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ProjectUserInvite'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Invite User to Project
      tags:
      - Project
  /project_users/{project_id}/permissions/{user_id}:
    get:
      description: Get the permissions of a user in a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get User Permissions for Project
      tags:
      - Project
...
  /user_projects/{owner_id}:
    get:
      description: Get all projects owned by a user
      parameters:
      - description: Owner User ID
        in: path
        name: owner_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Project'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Projects by Owner ID
      tags:
      - Project
  /user_shared_projects/{owner_id}:
    get:
      description: Get all projects owned by a user
      parameters:
      - description: Owner User ID
        in: path
        name: owner_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.JSONResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.Project'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.JSONResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.JSONResponse'
      summary: Get Projects by Owner ID
      tags:
      - Project
security:
- Authorization: []
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"